Changed contact tab to maps tab
