+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,13 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\src",
+    "\\src\\app",
+    "\\src\\pages",
+    "\\src\\pages\\home",
+    "\\src\\pages\\map",
+    "\\src\\pages\\tabs"
+  ],
+  "SelectedNode": "\\src\\pages\\home\\home.html",
+  "PreviewInSolutionExplorer": false
+}
