+++ config.xml
@@ -87,6 +87,6 @@
         <variable name="PLAY_SERVICES_VERSION" value="11.8.0" />
         <variable name="ANDROID_SUPPORT_V4_VERSION" value="24.1.0" />
     </plugin>
-    <engine name="android" spec="7.0.0" />
     <allow-navigation href="http://192.168.1.11:8100" />
+    <engine name="android" spec="~7.0.0" />
 </widget>
