+++ src/pages/.vs/VSWorkspaceState.json
@@ -3,6 +3,6 @@
     "",
     "\\home"
   ],
-  "SelectedNode": "\\home\\home.scss",
+  "SelectedNode": "\\home\\home.ts",
   "PreviewInSolutionExplorer": false
 }
