--- src/pages/contact/contact.html
@@ -1,11 +1,0 @@
-<ion-header>
-  <ion-navbar>
-    <ion-title>
-      Contact
-    </ion-title>
-  </ion-navbar>
-</ion-header>
-
-<ion-content>
-  <p>placeholder: put info about team here</p>
-</ion-content>
