--- src/pages/contact/contact.scss
@@ -1,3 +1,0 @@
-page-contact {
-
-}
