--- src/pages/contact/contact.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-import { NavController } from 'ionic-angular';
-
-@Component({
-  selector: 'page-contact',
-  templateUrl: 'contact.html'
-})
-export class ContactPage {
-
-  constructor(public navCtrl: NavController) {
-
-  }
-
-}
