+++ src/pages/home/home.html
@@ -7,8 +7,6 @@
 </ion-header>
 
 <ion-content padding>
-  <!--Create div to generate map inside-->
-  <div #map id="map" style="height:100%"></div>
   <!--Display scanner button and call scanner function in home.ts when clicked-->
   <div style="text-align:center; display:block; margin:1px;">
     <button id="scanButton" (click)='scanBarcode()' block>Scan</button>
