+++ src/pages/map/map.html
@@ -1,0 +1,12 @@
+<ion-header>
+  <ion-navbar>
+    <ion-title>
+      Map
+    </ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content>
+  <!--Create div to generate map inside-->
+  <div #map id="map" style="height:100%"></div>
+</ion-content>
