+++ src/pages/map/map.scss
@@ -1,0 +1,3 @@
+page-contact {
+
+}
