+++ src/pages/map/map.ts
@@ -1,0 +1,67 @@
+import { Component } from '@angular/core';
+import { NavController } from 'ionic-angular';
+
+// importing plugins
+import { GoogleMaps, GoogleMap, GoogleMapsEvent, GoogleMapOptions, CameraPosition, MarkerOptions, Marker } from '@ionic-native/google-maps';
+
+
+@Component({
+  selector: 'page-map',
+  templateUrl: 'map.html'
+})
+export class MapPage {
+
+  constructor(public navCtrl: NavController) {
+
+  }
+
+  // map plugin code
+  map: GoogleMap;
+
+  ionViewDidLoad() {
+   this.loadMap();
+  }
+
+  loadMap() {
+
+    // starting map options
+    let mapOptions: GoogleMapOptions = {
+      camera: {
+        target: {
+          lat: 43.0741904,
+          lng: -89.3809802
+        },
+        zoom: 18,
+        tilt: 30
+      }
+    };
+
+    this.map = GoogleMaps.create('map', mapOptions);
+
+    // Wait the MAP_READY before using any methods.
+    this.map.one(GoogleMapsEvent.MAP_READY)
+      .then(() => {
+        console.log('Map is ready!');
+
+        // Now you can use all methods safely.
+        // Add a marker to the map
+        this.map.addMarker({
+            title: 'Ionic',
+            icon: 'blue',
+            animation: 'DROP',
+            position: {
+              lat: 43.0741904,
+              lng: -89.3809802
+            }
+          })
+          // when marker is clicked, create popup alert
+          .then(marker => {
+            marker.on(GoogleMapsEvent.MARKER_CLICK)
+              .subscribe(() => {
+                alert('clicked');
+              });
+          });
+
+      });
+  }
+}
