+++ src/pages/tabs/tabs.html
@@ -1,5 +1,5 @@
 <ion-tabs selectedIndex="1">
   <ion-tab [root]="tab1Root" tabTitle="About" tabIcon="information-circle"></ion-tab>
   <ion-tab [root]="tab2Root" tabTitle="Scan" tabIcon="home"></ion-tab>
-  <ion-tab [root]="tab3Root" tabTitle="Contact" tabIcon="contacts"></ion-tab>
+  <ion-tab [root]="tab3Root" tabTitle="Map" tabIcon="map"></ion-tab>
 </ion-tabs>
