+++ src/pages/tabs/tabs.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 import { AboutPage } from '../about/about';
-import { ContactPage } from '../contact/contact';
+import { MapPage } from '../map/map';
 import { HomePage } from '../home/home';
 
 @Component({
@@ -11,7 +11,7 @@ export class TabsPage {
 
   tab1Root = AboutPage;
   tab2Root = HomePage;
-  tab3Root = ContactPage;
+  tab3Root = MapPage;
 
   constructor() {
